- Changed rounds to u8 instead of u32.

+++ src/sha/sha.rs
@@ -12,7 +12,7 @@ use crate::sha::structs::{HashError, HashFunction, HashResult, Word};
 fn process_block<T: Word + Default>(
 	mut state: Vec<T>,
 	block: &[u8],
-	rounds: u32,
+	rounds: u8,
 	k: &[T],
 ) -> Vec<T> {
 	let word_size = size_of::<T>();
@@ -85,7 +85,7 @@ fn process_block<T: Word + Default>(
 /// println!("My hash is {:?}!", hashed_sha256.unwrap());
 /// // Panics with HashError::TooManyRounds!
 /// ```
-pub fn sha(message: &str, hash_function: HashFunction, rounds: u32) -> Result<HashResult, HashError> {
+pub fn sha(message: &str, hash_function: HashFunction, rounds: u8) -> Result<HashResult, HashError> {
 	// Validate rounds
 	let max_rounds = hash_function.max_rounds();
 	if rounds > max_rounds {
